Fix FLEX yyleng type definition on 64-bit platforms.

+++ src/Headers/globals.h
@@ -35,7 +35,7 @@ extern /*@owned@*/ fileloc g_currentloc;
 /*@-declundef@*/ /* Might not process grammar files */
 extern /*@dependent@*/ /*@open@*/ FILE *yyin;
 extern /*@dependent@*/ /*@open@*/ /*@unused@*/ FILE *yyout;
-extern /*@unused@*/ int yyleng;
+extern /*@unused@*/ size_t yyleng;
 /*@=incondefs@*/ 
 
 extern int yydebug;
